requirements.txt
pytest==5.3.2
pytest-cov==2.8.1
pylint==2.4.4
flake8==3.7.9
bandit==1.6.2
kafka-python==2.0.1
confluent-kafka==1.3.0